fixed polydex
