+++ src/render/mod.rs
@@ -5,7 +5,6 @@ mod menu;
 pub(crate) mod message;
 mod palette;
 pub(crate) mod pipeline;
-mod polyblade;
 mod polydex;
 mod runner;
 mod state;
